+++ dart/lib/src/span_data_convention.dart
@@ -6,5 +6,10 @@ class SpanDataConvention {
   static const frozenFrames = 'frames.frozen';
   static const framesDelay = 'frames.delay';
 
+  // Thread/Isolate data keys according to Sentry span data conventions
+  // https://develop.sentry.dev/sdk/telemetry/traces/span-data-conventions/#thread
+  static const threadId = 'thread.id';
+  static const threadName = 'thread.name';
+
   // TODO: eventually add other data keys here as well
 }