+++ dart/test/sentry_client_lifecycle_test.dart
@@ -43,7 +43,8 @@ void main() {
         final client = fixture.getSut();
         fixture.options.logBatcher = MockLogBatcher();
 
-        client.lifeCycleRegistry.registerCallback<OnBeforeCaptureLog>((event) {
+        fixture.options.lifecycleRegistry
+            .registerCallback<OnBeforeCaptureLog>((event) {
           event.log.attributes['test'] =
               SentryLogAttribute.string('test-value');
         });
@@ -74,7 +75,8 @@ void main() {
         final client = fixture.getSut();
         fixture.options.logBatcher = MockLogBatcher();
 
-        client.lifeCycleRegistry.registerCallback<OnBeforeSendEvent>((event) {
+        fixture.options.lifecycleRegistry
+            .registerCallback<OnBeforeSendEvent>((event) {
           event.event.release = '999';
         });